+++ Tests/Opc.Ua.Core.Tests/Stack/State/NodeStateHandlerConcurrencyTests.cs
@@ -297,8 +297,8 @@ namespace Opc.Ua.Core.Tests.Stack.State
 
             testNodeState.Create(
                 systemContext,
-                new NodeId("TestNode", (ushort)7),
-                new QualifiedName("TestNode", (ushort)7),
+                new NodeId("TestNode", 7),
+                new QualifiedName("TestNode", 7),
                 new LocalizedText("TestNode"),
                 true);
 
@@ -325,8 +325,8 @@ namespace Opc.Ua.Core.Tests.Stack.State
 
             testNodeState.Create(
                 systemContext,
-                new NodeId("TestNode", (ushort)7),
-                new QualifiedName("TestNode", (ushort)7),
+                new NodeId("TestNode", 7),
+                new QualifiedName("TestNode", 7),
                 new LocalizedText("TestNode"),
                 true);
 
@@ -353,8 +353,8 @@ namespace Opc.Ua.Core.Tests.Stack.State
 
             testNodeState.Create(
                 systemContext,
-                new NodeId("TestNode", (ushort)7),
-                new QualifiedName("TestNode", (ushort)7),
+                new NodeId("TestNode", 7),
+                new QualifiedName("TestNode", 7),
                 new LocalizedText("TestNode"),
                 true);
 
@@ -381,8 +381,8 @@ namespace Opc.Ua.Core.Tests.Stack.State
 
             testNodeState.Create(
                 systemContext,
-                new NodeId("TestNode", (ushort)7),
-                new QualifiedName("TestNode", (ushort)7),
+                new NodeId("TestNode", 7),
+                new QualifiedName("TestNode", 7),
                 new LocalizedText("TestNode"),
                 true);
 
@@ -409,8 +409,8 @@ namespace Opc.Ua.Core.Tests.Stack.State
 
             testNodeState.Create(
                 systemContext,
-                new NodeId("TestNode", (ushort)7),
-                new QualifiedName("TestNode", (ushort)7),
+                new NodeId("TestNode", 7),
+                new QualifiedName("TestNode", 7),
                 new LocalizedText("TestNode"),
                 true);
 
@@ -437,8 +437,8 @@ namespace Opc.Ua.Core.Tests.Stack.State
 
             testNodeState.Create(
                 systemContext,
-                new NodeId("TestNode", (ushort)7),
-                new QualifiedName("TestNode", (ushort)7),
+                new NodeId("TestNode", 7),
+                new QualifiedName("TestNode", 7),
                 new LocalizedText("TestNode"),
                 true);
